+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master
 - Removed constant ParsleyUI
 
+- Parsley now relies on `input` events instead of `change` and `keyup` events.
+
 ## 2.2.0
 
 - type="number" now follows HTML5 spec. In particular, commas are no longer accepted. (#1037)